Move response classes from global/response to dto/common
