+++ sharedPrompts/src/main/java/org/example/sharedprompts/dto/report/validator/ValidReportProcessStatus.java
