--- sharedPrompts/src/main/java/org/example/sharedprompts/global/jwt/SecurityConfig.java
@@ -1,92 +1,0 @@
-package org.example.sharedprompts.global.jwt;
-
-import lombok.RequiredArgsConstructor;
-import org.example.sharedprompts.global.jwt.handler.OAuth2FailureHandler;
-import org.example.sharedprompts.global.jwt.handler.OAuth2SuccessHandler;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.web.cors.CorsConfigurationSource;
-
-@Configuration
-@RequiredArgsConstructor
-public class SecurityConfig {
-
-    private final JwtAuthFilter jwtAuthFilter;
-    private final OAuth2SuccessHandler oAuth2SuccessHandler;
-    private final OAuth2FailureHandler oAuth2FailureHandler;
-    private final CustomOAuth2UserService customOAuth2UserService;
-    private final JwtAuthenticationEntryPoint jwtAuthenticationEntryPoint;
-    private final CorsConfigurationSource corsConfigurationSource;
-
-    /* =========================
-       Password / Auth Manager
-       ========================= */
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    @Bean
-    public AuthenticationManager authenticationManager(
-            AuthenticationConfiguration configuration) throws Exception {
-        return configuration.getAuthenticationManager();
-    }
-
-    /* =========================
-       Security Filter Chain
-       ========================= */
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http
-                // CSRF, Session 사용 안 함
-                .csrf(AbstractHttpConfigurer::disable)
-                .sessionManagement(session ->
-                        session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
-
-                // CORS 설정 (Security Filter Chain 단일 처리)
-                .cors(cors -> cors.configurationSource(corsConfigurationSource))
-
-                // 인증 예외 처리
-                .exceptionHandling(exception ->
-                        exception.authenticationEntryPoint(jwtAuthenticationEntryPoint))
-
-                // 요청 허용/차단
-                .authorizeHttpRequests(auth -> auth
-                        .requestMatchers(
-                                "/auth/signup","/api/auth/signup",
-                                "/auth/login","/api/auth/login",
-                                "/auth/callback","/api/auth/callback",
-                                "/auth/refresh","/api/auth/refresh",
-                                "/oauth2/**", "/login/**"
-                        ).permitAll()
-                        .anyRequest().authenticated()
-                )
-
-                // OAuth2 로그인
-                .oauth2Login(oauth2 -> oauth2
-                        .userInfoEndpoint(userInfo ->
-                                userInfo.userService(customOAuth2UserService))
-                        .successHandler(oAuth2SuccessHandler)
-                        .failureHandler(oAuth2FailureHandler)
-                )
-
-                // JWT 인증 필터
-                .addFilterBefore(
-                        jwtAuthFilter,
-                        UsernamePasswordAuthenticationFilter.class
-                );
-
-        return http.build();
-    }
-}
