--- sharedPrompts/src/main/java/org/example/sharedprompts/global/jwt/TokenTtlProperties.java
@@ -1,22 +1,0 @@
-package org.example.sharedprompts.global.jwt;
-
-import lombok.Getter;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-@Getter
-@Component
-public class TokenTtlProperties {
-
-    private final long accessTokenValidity;
-    private final long refreshTokenValidity;
-
-    public TokenTtlProperties(
-            @Value("${jwt.access.expiration}") long accessTokenValidity,
-            @Value("${jwt.refresh.expiration}") long refreshTokenValidity
-    ) {
-        this.accessTokenValidity = accessTokenValidity;
-        this.refreshTokenValidity = refreshTokenValidity;
-    }
-
-}
